backend/requirements.txt
asgiref==3.8.1
boto3==1.35.72
botocore==1.35.72
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
Django==5.1.2
django-allauth==65.2.0
django-cors-headers==4.6.0
django-storages==1.14.4
djangorestframework==3.15.2
drf-yasg==1.21.8
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
jmespath==1.0.1
packaging==24.2
pillow==11.0.0
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
s3transfer==0.10.4
six==1.16.0
sqlparse==0.5.1
uritemplate==4.1.1
urllib3==2.2.3
python-dotenv==1.0.1
